+++ src/panoptes/data/observations.py
@@ -1,33 +1,17 @@
 import shutil
 import warnings
-from enum import IntEnum, auto
 from pathlib import Path
 
 import pandas as pd
 from astropy.nddata import CCDData, Cutout2D
 from astropy.utils.data import download_file
 from astropy.wcs import FITSFixedWarning
-from panoptes.utils.images import fits as fits_utils
 from tqdm.auto import tqdm
 
 from panoptes.data.settings import CloudSettings
+from panoptes.utils.images import fits as fits_utils
 
 warnings.filterwarnings('ignore', category=FITSFixedWarning)
-
-
-class ObservationStatus(IntEnum):
-    """The status of an observation."""
-    ERROR = auto()
-    NOT_ENOUGH_FRAMES = auto()
-    UNKNOWN = auto()
-    CREATED = auto()
-    RECEIVING = auto()
-    RECEIVED = auto()
-    PROCESSING = auto()
-    CALIBRATING = auto()
-    CALIBRATED = auto()
-    MATCHING = auto()
-    MATCHED = auto()
 
 
 class ObservationInfo: