+++ src/panoptes/data/utils/cli/main.py
@@ -1,21 +1,20 @@
 from pathlib import Path
-from typing import List, Union
 
 import pandas as pd
 import typer
 from astropy.time import Time
-from panoptes.utils.time import current_time, flatten_time
 from rich import print
 from tqdm import tqdm
 
 from panoptes.data.observations import ObservationInfo
 from panoptes.data.search import search_observations
+from panoptes.utils.time import current_time, flatten_time
 
 app = typer.Typer(pretty_exceptions_enable=False)
 
 
 @app.command()
-def download(sequence_id: Union[str, None] = typer.Argument(..., help='Sequence ID for the Observation.'),
+def download(sequence_id: str | None = typer.Argument(..., help='Sequence ID for the Observation.'),
              output_dir: Path = typer.Option(
                  None,
                  '--output-dir', '-o',
@@ -26,7 +25,7 @@ def download(sequence_id: Union[str, None] = typer.Argument(..., help='Sequence 
                  '--image-query', '-q',
                  help='Query for images, default \'status!="ERROR"\''
              ),
-             ) -> List[str]:
+             ) -> list[str]:
     """Downloads all FITS images for the observation."""
     local_files = list()
 
@@ -49,20 +48,20 @@ def download(sequence_id: Union[str, None] = typer.Argument(..., help='Sequence 
 
 @app.command()
 def get_metadata(
-    sequence_id: Union[str, None] = typer.Option(
+    sequence_id: str | None = typer.Option(
         None, '--sequence-id', '-s',
         help='Sequence ID for the Observation.'
     ),
-    unit_id: Union[str, None] = typer.Option(
+    unit_id: str | None = typer.Option(
         None, '--unit-id', '-u',
         help='Unit ID for the Observation. '
              'Use with a date range to download all metadata for a unit.'
     ),
-    start_date: Union[str, None] = typer.Option(
+    start_date: str | None = typer.Option(
         None, '--start-date', '-s',
         help='Start date for downloading metadata in form YYYY-MM-DD'
     ),
-    end_date: Union[str, None] = typer.Option(
+    end_date: str | None = typer.Option(
         None, '--end-date', '-s',
         help='End date for downloading metadata, defaults to now.'
     ),
@@ -115,7 +114,7 @@ def get_metadata(
                 try:
                     pbar.set_description(f'Getting metadata for {rec["sequence_id"]}')
                     dfs.append(ObservationInfo(meta=rec).image_metadata)
-                except Exception as e:
+                except Exception:
                     pbar.write(f'Error in {idx} {rec["sequence_id"]}')
 
             pd.concat(dfs).to_csv(output_fn)