+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## UNRELEASED
+
+- Modernize the repo to use `pyproject.toml`.
+- Update docs.
+- Add basic tests.
+- Releases are created directly by GitHub Actions.
+
+## Version 0.2.0 (2024-10-08)
+
+- Print length of observations.
+- Only apply status filter if given.
+
+## Version 0.1.9 (2024-10-07)
+
+- Do better path matching for image url for new scheme.
+
+## Version 0.1.8 (2024-10-07)
+
+- Merge branch 'main' of github.com:panoptes/panoptes-data
+
+## Version 0.1.7 (2024-04-18)
+
+- Fix image_list for latest processing.
+
+## Version 0.1.6 (2024-04-18)
+
+- Change default bucket for downloading images and also make it a parameter.
+
+## Version 0.1.5 (2024-02-02)
+
+- Fixing release for pypi api tokens.
+
+## Version 0.1.4 (2024-02-02)
+
+- Add GitHub Actions PyPI release workflow.
+
 ## Version 0.1.3
 
 - Cleanup of images and observations.
@@ -44,7 +80,15 @@
 - Added basic cli interface for download images and metadata for observations.
 - Fixed install dependencies.
 
+## Version 0.0.3 (2022-06-24)
+
+- More cleanup of dependencies for release.
+
 ## Version 0.0.2
 
 - Observation search available via ``panoptes.data.search.search_observations``.
 - ``ObservationInfo`` for working with observation data and metadata.
+
+## Version 0.0.1 (2022-06-23)
+
+- Fixing setup.